requirements.txt
mysql.connector
fastapi
uvicorn